+++ stdnet/lib/redis/client.py
@@ -30,12 +30,15 @@ redis_command = namedtuple('redis_command','command args options callbacks')
 
 __all__ = ['Redis',
            'RedisProxy',
-           'Pipeline',
-           ]
+           'PrefixedRedis',
+           'Pipeline']
 
 
 collection_list = (tuple, list, set, frozenset, dict)
 
+
+def raise_error(exception=NotImplementedError):
+    raise exception()
 
 def list_or_args(keys, args = None):
     if not isinstance(keys, collection_list):
@@ -222,7 +225,6 @@ class Redis(object):
         'EVALSHA': eval_command_callback,
         'EVAL': eval_command_callback
     }
-
     _STATUS = ''
 
     def __init__(self, connection_pool=None, **connection_kwargs):
@@ -240,6 +242,10 @@ class Redis(object):
     def pipelined(self):
         return False
     
+    @property
+    def prefix(self):
+        return ''
+        
     @property
     def encoding(self):
         return self.connection_pool.encoding
@@ -266,12 +272,21 @@ between the client and server.
         return Pipeline(self)
     
     def prefixed(self, prefix):
+        '''Return a new :class:`PrefixedRedis` client'''
         return PrefixedRedis(self, prefix)
 
+    def preprocess_command(self, cmnd, *args, **options):
+        return args, options
+        
+    def request(self, cmnd, *args, **options):
+        '''Return a new :class:`RedisRequest`'''
+        connection = self.connection_pool.get_connection()
+        args, options = self.preprocess_command(cmnd, *args, **options)
+        return connection.request(self, cmnd, *args, **options)
+    
     def execute_command(self, *args, **options):
         "Execute a command and return a parsed response"
-        connection = self.connection_pool.get_connection()
-        return connection.execute_command(self, *args, **options)
+        return self.request(*args, **options).execute()
 
     def _parse_response(self, request, response, command_name, args, options):
         callbacks = self.response_errbacks if isinstance(response, Exception)\
@@ -320,7 +335,7 @@ between the client and server.
         return self.execute_command('DEL', *names)
     __delitem__ = delete
 
-    def flushall(self):
+    def flushall(self): #pragma    nocover
         "Delete all keys in all databases on the current host"
         return self.execute_command('FLUSHALL')
 
@@ -354,7 +369,7 @@ between the client and server.
         """
         return self.execute_command('SAVE')
 
-    def shutdown(self):
+    def shutdown(self): #pragma    nocover
         "Shutdown the server"
         try:
             self.execute_command('SHUTDOWN')
@@ -851,28 +866,26 @@ The first element is the score and the second is the value.'''
 
     # zset script commands
 
-    def zpopbyrank(self, name, start, stop = None, withscores = False,
-                   desc = False, **options):
+    def zpopbyrank(self, name, start, stop=None, withscores=False,
+                   desc=False, **options):
         '''Pop a range by rank'''
         options['withscores'] = withscores
         stop = stop if stop is not None else start
-        return self.script_call('zpop', name,
-                                'rank',
+        return self.script_call('zpop', (name,), 'rank',
                                 start, stop, int(desc), int(withscores),
                                 **options)
 
-    def zpopbyscore(self, name, start, stop = None, withscores = False,
-                    desc = False, **options):
+    def zpopbyscore(self, name, start, stop=None, withscores=False,
+                    desc=False, **options):
         '''Pop a range by score'''
         options['withscores'] = withscores
         stop = stop if stop is not None else start
-        return self.script_call('zpop', name,
-                                'score',
+        return self.script_call('zpop', (name,), 'score',
                                 start, stop, int(desc), int(withscores),
                                 **options)
 
     def _zaggregate(self, command, dest, keys,
-                    aggregate=None, withscores = None, **options):
+                    aggregate=None, withscores=None, **options):
         pieces = [command, dest, len(keys)]
         if isinstance(keys, dict):
             items = keys.items()
@@ -979,7 +992,7 @@ The first element is the score and the second is the value.'''
         '''Execute a registered lua script.'''
         script = get_script(name)
         if not script:
-            raise ValueError('No such script {0}'.format(name))
+            raise RedisError('No such script {0}'.format(name))
         return script.evalsha(self, keys, *args, **options)
 
     def script_flush(self):
@@ -1002,14 +1015,20 @@ The first element is the score and the second is the value.'''
 
 
 class RedisProxy(Redis):
+    '''A proxy to a :class:`Redis` client. It is the base class
+of :class:`PrefixedRedis` and :class:`Pipeline`.
 
+.. attribute:: client
+
+    The underlying :class:`Redis` client
+'''
     def __init__(self, client):
         self.__client = client
 
     @property
     def client(self):
         return self.__client
-
+    
     @property
     def connection_pool(self):
         return self.client.connection_pool
@@ -1026,35 +1045,151 @@ class RedisProxy(Redis):
     def encoding(self):
         return self.client.encoding
 
+    def clone(self, **kwargs):
+        c = copy(self)
+        c.__client = c.client.clone(**kwargs)
+        return c
+        
+
+prefix_all = lambda pfix, args: ['%s%s' % (pfix, a) for a in args]
+prefix_alternate = lambda pfix, args: [a if n//2*2==n else '%s%s' % (pfix, a)\
+                                       for n, a in enumerate(args,1)]
+prefix_not_last = lambda pfix, args: ['%s%s' % (pfix, a) for a in args[:-1]]\
+                                        + [args[-1]]
+prefix_not_first = lambda pfix, args: [args[0]] +\
+                                      ['%s%s' % (pfix, a) for a in args[1:]]
+
+def prefix_zinter(pfix, args):
+    dest, numkeys, params = args[0], args[1], args[2:]
+    args = ['%s%s' % (pfix, dest), numkeys]
+    nk = 0
+    for p in params:
+        if nk < numkeys:
+            nk += 1
+            p = '%s%s' % (pfix, p)
+        args.append(p)
+    return args
+
+def prefix_sort(pfix, args):
+    prefix = True
+    nargs = []
+    for a in args:
+        if prefix:
+            a = '%s%s' % (pfix, a)
+            prefix = False
+        elif a in ('BY', 'GET', 'STORE'):
+            prefix = True
+        nargs.append(a)
+    return nargs
+    
+def pop_list_result(pfix, result):
+    if result:
+        return (result[0][len(pfix):], result[1])
+
 
 class PrefixedRedis(RedisProxy):
-    '''A prefixed redis client. It append a prefix to all keys.'''
+    '''A :class:`RedisProxy` for a prefixed redis client.
+It append a prefix to all keys.
+
+.. attribute:: prefix
+
+    The prefix for this :class:`PrefixedRedis`
     
-    SERVER_COMMANDS = frozenset(('BGREWRITEOF', 'BGSAVE', 'CLIENT', 'CONFIG',
-                                 'DBSIZE', 'DEBUG', 'SHUTDOWN', 'SLAVEOF',
-                                 'SLOWLOG','SYNC', 'TIME'))
-    def __init__(self, prefix, client):
-        self.prefix = prefix
+Typical usage::
+
+    >>> from stdnet import getdb
+    >>> redis = getdb().client
+    >>> pr = redis.prefixed('myprefix.')
+    >>> pr.set('a','foo')
+    True
+    >>> pr.get('a')
+    'foo'
+    >>> redis.get('a')
+    None
+    >>> redis.get('myprefix.a')
+    'foo'
+'''    
+    EXCLUDE_COMMANDS = frozenset(('BGREWRITEOF', 'BGSAVE', 'CLIENT', 'CONFIG',
+                                  'DBSIZE', 'DEBUG', 'DISCARD', 'ECHO',
+                                  'EVAL', 'EVALSHA', 'EXEC',
+                                  'INFO', 'LASTSAVE', 'PING',
+                                  'PSUBSCRIBE', 'PUBLISH', 'PUNSUBSCRIBE',
+                                  'QUIT', 'RANDOMKEY', 'SAVE', 'SCRIPT',
+                                  'SELECT', 'SHUTDOWN', 'SLAVEOF', 
+                                  'SLOWLOG', 'SUBSCRIBE', 'SYNC',
+                                  'TIME', 'UNSUBSCRIBE', 'UNWATCH'))
+    SPECIAL_COMMANDS = {
+        'BITOP': prefix_not_first,
+        'BLPOP': prefix_not_last,
+        'BRPOP': prefix_not_last,
+        'BRPOPLPUSH': prefix_not_last,
+        'RPOPLPUSH': prefix_all,
+        'DEL': prefix_all,
+        'FLUSHDB': lambda prefix, args: raise_error(),
+        'FLUSHALL': lambda prefix, args: raise_error(),
+        'MGET': prefix_all,
+        'MSET': prefix_alternate,
+        'MSETNX': prefix_alternate,
+        'MIGRATE': prefix_all,
+        'RENAME': prefix_all,
+        'RENAMENX': prefix_all,
+        'SDIFF': prefix_all,
+        'SDIFFSTORE': prefix_all,
+        'SINTER': prefix_all,
+        'SINTERSTORE': prefix_all,
+        'SMOVE': prefix_not_last,
+        'SORT': prefix_sort,
+        'SUNION': prefix_all,
+        'SUNIONSTORE': prefix_all,
+        'WATCH': prefix_all,
+        'ZINTERSTORE': prefix_zinter,
+        'ZUNIONSTORE': prefix_zinter
+    }
+    RESPONSE_CALLBACKS = {
+        'KEYS': lambda pfix, response: [r[len(pfix):] for r in response],
+        'BLPOP': pop_list_result,
+        'BRPOP': pop_list_result
+    }
+    def __init__(self, client, prefix):
         super(PrefixedRedis, self).__init__(client)
+        self.__prefix = prefix
         
-    def execute_command(self, cmnd, *args, **options):
-        if args and cmnd not in self.SERVER_COMMANDS:
-            args = list(args)
-            args[0] = '%s%s' % (self.prefix, args[0])
-        return self.client.execute_command(cmnd, *args, **options)
+    @property
+    def prefix(self):
+        return self.__prefix
     
+    def preprocess_command(self, cmnd, *args, **options):
+        if args and cmnd not in self.EXCLUDE_COMMANDS:
+            handle = self.SPECIAL_COMMANDS.get(cmnd, self.handle)
+            args = handle(self.prefix, args)
+        return args, options
+    
+    def handle(self, prefix, args):
+        if args:
+            args = list(args)
+            args[0] = '%s%s' % (prefix, args[0])
+        return args
+        
     def dbsize(self):
         return self.client.countpattern('%s*' % self.prefix)
     
     def flushdb(self):
         return self.client.delpattern('%s*' % self.prefix)
     
-    def delpattern(self, pattern):
-        return self.client.delpattern('%s%s' % (self.prefix, pattern))
-    
-    def flushall(self):
-        raise NotImplementedError()
+    def _eval(self, command, body, keys, *args, **options):
+        if keys:
+            keys = ['%s%s' % (self.prefix, k) for k in keys]
+        return super(PrefixedRedis, self)._eval(command, body, keys, *args,
+                                                **options)
         
+    def _parse_response(self, request, response, command_name, args, options):
+        if command_name in self.RESPONSE_CALLBACKS:
+            if not isinstance(response, Exception):
+                response = self.RESPONSE_CALLBACKS[command_name](self.prefix,
+                                                                 response)
+        return self.client._parse_response(request, response, command_name,
+                                           args, options)
+    
         
 class Pipeline(RedisProxy):
     """A :class:`Pipeline` provide a way to commit multiple commands
@@ -1107,11 +1242,9 @@ At some other point, you can then run: pipe.execute(),
 which will execute all commands queued in the pipe.
 """
         callbacks = []
-        #callback = self.client.response_callbacks.get(cmnd)
-        #if callback:
-        #    callbacks.append(callback)
+        args, options = self.client.preprocess_command(cmnd, *args, **options)
         self.command_stack.append(
-                        redis_command(cmnd, args, options, callbacks))
+                        redis_command(cmnd, tuple(args), options, callbacks))
         return self
 
     def add_callback(self, callback):
@@ -1134,19 +1267,17 @@ Several callbacks can be added for a given command::
     pipe.sadd('foo').add_callback(mycallback).add_callback(mycallback2)
 '''
         if self.empty:
-            raise ValueError('Cannot add callback. No command in the stack')
+            raise RedisError('Cannot add callback. No command in the stack')
         self.command_stack[-1].callbacks.append(callback)
         return self
 
     def parse_response(self, request):
-        response = request.response
-        commands = request.args
-        processed = []
-        response = response[-1]
-        commands = commands[1:-1]
+        response = request.response[-1]
+        commands = request.args[1:-1]
         if len(response) != len(commands):
             raise ResponseError("Wrong number of response items from "
                                 "pipeline execution")
+        processed = []
         parse_response = self._parse_response
         for r, cmd in zip(response, commands):
             command, args, options, callbacks = cmd
@@ -1154,23 +1285,20 @@ Several callbacks can be added for a given command::
             for callback in callbacks:
                 r = callback(processed, r)
             processed.append(r)
-        return processed
+        if request.load_script:
+            return load_missing_scripts(self, commands, processed)
+        else:
+            return processed
 
-    def execute(self, load_script=False):
-        '''Execute all commands in the current pipeline.'''
+    def request(self, load_script=False):
         self.execute_command('EXEC')
         commands = self.command_stack
         self.reset()
         conn = self.connection_pool.get_connection()
-        res = conn.execute_pipeline(self, commands)
-        if isinstance(res, RedisRequest):
-            return res.add_callback(partial(self.finalise,commands,load_script))
-        else:
-            return self.finalise(commands, load_script, res)
-
-    def finalise(self, commands, load_script, result):
-        if load_script:
-            return load_missing_scripts(self, commands[1:-1], result)
-        return result
-
-
+        request = conn.request(self, None, *commands)
+        request.load_script = load_script
+        return request
+        
+    def execute(self, load_script=False):
+        '''Execute all commands in the current pipeline.'''
+        return self.request(load_script).execute()