+++ tests/regression/pubsub.py
@@ -26,8 +26,9 @@ class TestPubSub(test.TestCase):
         self.p = self.publisher()
     
     def tearDown(self):
-        self.s.unsubscribe()
-        self.s.punsubscribe()
+        s = self.s
+        self.s = None
+        s.disconnect()
         
     def testDummy(self):
         p = DummyBackendDataServer('','')
@@ -47,4 +48,22 @@ class TestPubSub(test.TestCase):
         self.assertEqual(len(res), 1)
         data = s.get_all()
         self.assertEqual(data['test'],['hello world!'])
+        s.unsubscribe()
+        res = list(s.pool(1))
+        self.assertFalse(s.subscription_count())
+        
+    def testPsubscribe(self):
+        p = self.p
+        s = self.s
+        self.assertEqual(s.psubscribe('test.*'), [b'psubscribe', b'test.*', 1])
+        self.assertEqual(s.subscription_count(), 1)
+        self.assertEqual(p.publish('test.bla', 'hello world!'), 1)
+        res = list(s.pool(1))
+        self.assertEqual(len(res), 1)
+        data = s.get_all('test.*')
+        self.assertEqual(data['test.bla'], ['hello world!'])
+        s.punsubscribe()
+        res = list(s.pool(1))
+        self.assertFalse(s.subscription_count())
+